Time: 4 ms (30.57%), Space: 8.6 MB (64.76%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        long rev = 0, rem;
+        long long rev = 0, rem;
         int temp = x;
         if(x < 0) {
             return false;
